refactor: refactor cjsToEsm.cjs to esm
